feat(navigation): implement workspace tool switcher and refactor header architecture
* Replace `VISearchHeader` with a modular `AppHeader` organization.
* Introduce `VWorkspaceBar` component for cyclic tool switching and selection.
* Update `VDropdownMenu` positioning for improved layout stability.
* Rename `ISearchPage` and `ISearchTemplate` to `MainPage` and `AppTemplate` to align with the new navigation structure.
* Update router configurations to support the unified header layout.
